requirements.txt
# Core dependencies
torch>=2.7.0
torchaudio>=2.7.0
transformers>=4.51.3
vllm>=0.9.0

# Audio processing
soundfile>=0.12.1
librosa>=0.10.0
pydub>=0.25.1

# CosyVoice dependencies
onnxruntime-gpu>=1.16.0
openai-whisper>=20231117
WeTextProcessing>=1.0.3
hyperpyyaml>=1.2.2
conformer>=0.3.2
lightning>=2.0.0
hydra-core>=1.3.2
HyperPyYAML>=1.2.2
modelscope>=1.9.5

# Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Monitoring & utilities
prometheus-client>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
PyYAML>=6.0.1
numpy>=1.24.0
tqdm>=4.66.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
